use pointer to User everywhere.
this was motivated by a supposed bug in which in the first call of someone to the bot the chat id
wouldn't be updated and we would try to respond to chat id 0.

could have been fixed by a hack but I think this was cleaner even though it touched the entire
codebase and may have introduced bugs.