do not delete as failed payments that were already marked as successful.
this can happen if we override a hosted channel on cliche that had a revealed-but-uncommitted preimage.

the reveal is already enough for lntxbot to mark it as paid, which is correct.
however after the override immortan/cliche will notify it as failed, which would cause lntxbot to delete it.

but we don't want to delete it.